Fix directory name and add get_valuation function
